+++ Remnants/Main.cs
@@ -17,7 +17,7 @@ namespace Remnants
         #region Variables
         private const string modGUID = "KawaiiBone.Remnants";
         private const string modName = "Remnants";
-        private const string modVersion = "1.2.15";
+        private const string modVersion = "1.2.16";
 
         public static Remnants Instance;
         private readonly Harmony _harmony = new Harmony(modGUID);